fixing bug and adding docs

+++ ciftify/qc_config.py
@@ -198,7 +198,7 @@ class Scene(QCScene):
                 half_the_height = height / 2
                 img_top = img.crop((0, 0, width, half_the_height))
                 img_btm = img.crop((0, half_the_height, width, half_the_height))
-                im2 = Image.new('RGBA', (width*2, half_the_height))
+                im2 = Image.new('RGBA', (int(width*2), half_the_height))
                 im2.paste(img_top, (0, 0))
                 im2.paste(img_bottom, (0, width))
                 im2.save(output_loc)
